requirements.txt
django  # Web framework
djangorestframework  # For building REST APIs
djangorestframework-simplejwt  # For JWT authentication
django-cors-headers  # For handling Cross-Origin Resource Sharing (CORS)
drf-yasg  # For API documentation
black  # Code formatter
django-environ #.env
Pillow #image